откорректирован функционал для комментариев, добавлены роли

+++ src/main/java/com/skypro/resale/repository/CommentRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
 
     Optional<Comment> findByIdAndAdsId(Integer commentId, Integer adId);
